netloc: new netloc_arch_t,  use MPI for resources
It is not possible anymore the represent the all machine in one
Scotch_arch. We use one arch for the nodes, and for each node we use
another arch for the slots.
The available resources are not got from resource manager, but by using
a new standalone program (netloc_mpi_find_hosts) using MPI and hwloc.