netloc: Encapsulate <nodes> and <edges> into a <explicit> tag in the XML file.
This encapsulation is used in order to add the logical topology to the
XML in the forthcoming versions.

Export, import and tests are updated accordingly.

Add a extra function to properly handle whether or not warning and
error messages should be displayed. The rule used so far is if debug,
default is print ; if not debug, default is not to print. In any case,
HWLOC_XML_VERBOSE can be use to specify the desired behaviour.

Signed-off-by: Clement Foyer <clement.foyer@inria.fr>